+++ examples/resilience_testing/lib/resilience_testing.ex
@@ -1,0 +1,18 @@
+defmodule ResilienceTesting do
+  @moduledoc """
+  Documentation for `ResilienceTesting`.
+  """
+
+  @doc """
+  Hello world.
+
+  ## Examples
+
+      iex> ResilienceTesting.hello()
+      :world
+
+  """
+  def hello do
+    :world
+  end
+end